+++ angular-workspace/projects/ni/nimble-angular/src/thirdparty/directives/radio_control_value_accessor.ts
@@ -1,10 +1,12 @@
 /**
  * [Nimble]
- * Copied from https://github.com/angular/angular/blob/14.2.6/packages/forms/src/directives/radio_control_value_accessor.ts
+ * Copied from https://github.com/angular/angular/blob/15.2.10/packages/forms/src/directives/radio_control_value_accessor.ts
  * with the following modifications:
  * - Changed throwNameError() to throw Error instead of RuntimeError. This makes the file compile with Angular version 12.
  * - Removed now-unused import for RuntimeErrorCode and RuntimeError
  * - Updated import of ControlValueAccessor, NgControl, and NG_VALUE_ACCESSOR to pull from package export
+ * - Remove all configuration from RadioControlValueAccessor's `@Directive` decorator
+ * - Remove use of internal CALL_SET_DISABLED_STATE injection token, and inline default value
  */
 
 /**
@@ -15,16 +17,18 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Directive, ElementRef, forwardRef, Injectable, Injector, Input, NgModule, OnDestroy, OnInit, Renderer2} from '@angular/core';
+import {Directive, ElementRef, Injectable, Injector, Input, NgModule, OnDestroy, OnInit, Renderer2} from '@angular/core';
 
 import {BuiltInControlValueAccessor} from './control_value_accessor';
-import {ControlValueAccessor, NgControl, NG_VALUE_ACCESSOR} from '@angular/forms';
+import {ControlValueAccessor, NgControl, SetDisabledStateOption} from '@angular/forms';
 
+/* [Nimble] Do not register as a value accessor provider
 export const RADIO_VALUE_ACCESSOR: any = {
   provide: NG_VALUE_ACCESSOR,
   useExisting: forwardRef(() => RadioControlValueAccessor),
   multi: true
 };
+*/
 
 function throwNameError() {
   /* [Nimble] RuntimeErrorCode is not exported from @angular/forms in version 12; falling back to version 12 behavior
@@ -117,12 +121,15 @@ export class RadioControlRegistry {
  * @ngModule FormsModule
  * @publicApi
  */
+/* [Nimble] Remove all configuration from @Directive decorator
 @Directive({
   selector:
       'input[type=radio][formControlName],input[type=radio][formControl],input[type=radio][ngModel]',
   host: {'(change)': 'onChange()', '(blur)': 'onTouched()'},
   providers: [RADIO_VALUE_ACCESSOR]
 })
+*/
+@Directive()
 export class RadioControlValueAccessor extends BuiltInControlValueAccessor implements
     ControlValueAccessor, OnDestroy, OnInit {
   /** @internal */
@@ -134,6 +141,8 @@ export class RadioControlValueAccessor extends BuiltInControlValueAccessor imple
   /** @internal */
   // TODO(issue/24571): remove '!'.
   _fn!: Function;
+
+  private setDisabledStateFired = false;
 
   /**
    * The registered callback function called when a change event occurs on the input element.
@@ -164,6 +173,10 @@ export class RadioControlValueAccessor extends BuiltInControlValueAccessor imple
    * Tracks the value of the radio input element
    */
   @Input() value: any;
+
+  // [Nimble]: Can't override default behavior by injection token, because it is not exported. Inlining value of setDisabledStateDefault.
+  private callSetDisabledState: SetDisabledStateOption = 'always';
+      //inject(CALL_SET_DISABLED_STATE, {optional: true}) ?? setDisabledStateDefault;
 
   constructor(
       renderer: Renderer2, elementRef: ElementRef, private _registry: RadioControlRegistry,
@@ -202,6 +215,33 @@ export class RadioControlValueAccessor extends BuiltInControlValueAccessor imple
       fn(this.value);
       this._registry.select(this);
     };
+  }
+
+  /** @nodoc */
+  override setDisabledState(isDisabled: boolean): void {
+    /**
+     * `setDisabledState` is supposed to be called whenever the disabled state of a control changes,
+     * including upon control creation. However, a longstanding bug caused the method to not fire
+     * when an *enabled* control was attached. This bug was fixed in v15 in #47576.
+     *
+     * This had a side effect: previously, it was possible to instantiate a reactive form control
+     * with `[attr.disabled]=true`, even though the the corresponding control was enabled in the
+     * model. This resulted in a mismatch between the model and the DOM. Now, because
+     * `setDisabledState` is always called, the value in the DOM will be immediately overwritten
+     * with the "correct" enabled value.
+     *
+     * However, the fix also created an exceptional case: radio buttons. Because Reactive Forms
+     * models the entire group of radio buttons as a single `FormControl`, there is no way to
+     * control the disabled state for individual radios, so they can no longer be configured as
+     * disabled. Thus, we keep the old behavior for radio buttons, so that `[attr.disabled]`
+     * continues to work. Specifically, we drop the first call to `setDisabledState` if `disabled`
+     * is `false`, and we are not in legacy mode.
+     */
+    if (this.setDisabledStateFired || isDisabled ||
+        this.callSetDisabledState === 'whenDisabledForLegacyCode') {
+      this.setProperty('disabled', isDisabled);
+    }
+    this.setDisabledStateFired = true;
   }
 
   /**