+++ packages/nimble-components/src/rich-text/models/markdown-parser-mention-configuration.ts
@@ -1,7 +1,4 @@
-import type {
-    MappingConfigs,
-    RichTextMentionConfig
-} from '../../rich-text-mention/base';
+import type { MappingConfigs } from '../../rich-text-mention/base/types';
 import type { MentionInternals } from '../../rich-text-mention/base/models/mention-internals';
 
 /**
@@ -14,12 +11,17 @@ export class MarkdownParserMentionConfiguration {
     private readonly regexPattern: RegExp;
     private readonly mappingConfigs?: MappingConfigs;
 
-    public constructor(
-        mentionInternals: MentionInternals<RichTextMentionConfig>
-    ) {
+    public constructor(mentionInternals: MentionInternals) {
         this.regexPattern = new RegExp(mentionInternals.pattern ?? '');
-        this.mappingConfigs = mentionInternals.mentionConfig?.mappingConfigs;
+        this.mappingConfigs = mentionInternals.mappingConfigs;
         this.viewElement = mentionInternals.viewElement;
+    }
+
+    public static isObservedMentionInternalsProperty(arg: unknown): boolean {
+        return (
+            typeof arg === 'string'
+            && ['pattern', 'mappingConfigs'].includes(arg)
+        );
     }
 
     public isValidMentionHref(mentionHref: string): boolean {